+++ cpp/benchmarks/copying/scatter_benchmark.cu
@@ -16,7 +16,7 @@
 
 #include <benchmark/benchmark.h>
 
-#include <cudf/copying.hpp>
+#include <cudf/legacy/copying.hpp>
 #include <cudf/legacy/table.hpp>
 
 #include <gmock/gmock.h>