+++ cpp/benchmarks/type_dispatcher/type_dispatcher_benchmark.cu
@@ -15,7 +15,7 @@
  */
 
 
-#include <cudf/copying.hpp>
+#include <cudf/legacy/copying.hpp>
 #include <cudf/legacy/table.hpp>
 
 #include <gmock/gmock.h>